refactor(JsonManager.java): Removed the JacksonSerializers class and moved the serializer/deserializ
